# docs/swagger.yaml
basePath: /
definitions:
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  models.Orders:
    properties:
      cargoWeight:
        type: integer
      clientId:
        type: integer
      date:
        type: string
      deliveryDate:
        type: string
      departure:
        type: string
      destination:
        type: string
      driverid:
        type: integer
      id:
        type: integer
      managerId:
        type: string
      price:
        type: integer
    required:
    - cargoWeight
    - clientId
    - deliveryDate
    - departure
    - destination
    - driverid
    - price
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for Delivery company Application
  license: {}
  title: Delivery company
  version: "1.0"
paths:
...
  /client-api/orders/:
    get:
      description: Get all orders by a client
      operationId: get-all-orders-client
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all client orders
      tags:
      - Client
  /client-api/orders/{id}:
    get:
      consumes:
      - application/json
      description: get order by id client
      operationId: get-order-byid-client
      parameters:
      - description: ordre id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get order by id
      tags:
      - Client
  /client-api/orders/active:
    get:
      description: Get active orders by a client
      operationId: get-all-active-orders-client
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get active orders
      tags:
      - Client
  /client-api/orders/completed:
    get:
      description: Get completed orders by a client
      operationId: get-all-completed-orders-client
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get completed orders
      tags:
      - Client
  /client-api/orders/search/{city}:
    get:
      consumes:
      - application/json
      description: search order by a client
      operationId: search-order-client
      parameters:
      - description: name city
        in: path
        name: city
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Search order
      tags:
      - Client
...
  /manager-api/orders/:
    get:
      description: Get all orders by a manager
      operationId: get-all-orders-manager
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all orders
      tags:
      - Manager
    post:
      consumes:
      - application/json
      description: creating an order by a manager
      operationId: creat-order-manager
      parameters:
      - description: order info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Orders'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Creat order
      tags:
      - Manager
  /manager-api/orders/{id}:
    delete:
      consumes:
      - application/json
      description: delete order by a manager
      operationId: delete-order-manager
      parameters:
      - description: order id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete order
      tags:
      - Manager
    get:
      consumes:
      - application/json
      description: get order by id manager
      operationId: get-order-byid-manager
      parameters:
      - description: ordre id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get order by id
      tags:
      - Manager
    put:
      consumes:
      - application/json
      description: update order by a manager
      operationId: update-order-manager
      parameters:
      - description: order id
        in: path
        name: id
        required: true
        type: integer
      - description: new info order
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Orders'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update order
      tags:
      - Manager
  /manager-api/orders/active:
    get:
      description: Get all active orders by a manager
      operationId: get-all-active-orders-manager
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all active orders
      tags:
      - Manager
  /manager-api/orders/completed:
    get:
      description: Get all completed orders by a manager
      operationId: get-all-completed-orders-manager
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all completed orders
      tags:
      - Manager
  /manager-api/orders/search/{city}:
    get:
      consumes:
      - application/json
      description: search order by a manager
      operationId: search-order-manager
      parameters:
      - description: name city
        in: path
        name: city
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Search order
      tags:
      - Manager
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"